Isolate the aws-java-sdk-s3 dependency to S3MockRule
There's a ton of different versions of the Java S3 SDK [1} . A version
is used at the client side is most certainly will not match with
the version declared in S3Mock. In case if the library integrated to a
Java application via Maven/Gradle, the developer will have to exclude
the aws-java-sdk-s3 dependency or to be at mercy of the default version
resolution process in the dependency manager.

It makes sense to declare the dependency as optional and isolate
it to `S3MockRule`. In this case the dependency manager will pick up
the version of the client SDK. In case the server is run as a Docker
container, we will not add the library to the a distribution, because
the server doesn't depend on it.

[1]: https://mvnrepository.com/artifact/com.amazonaws/aws-java-sdk-s3