Remove handling of v4 signing
Removal of the handling of v4 signing fixes a bug that was already
in `FileStore.copyPart` before: if v4 signing would have been "requested"
(according to the request headers), this could have caused a `BufferOverflowException`:
with v4 signing the file input stream was wrapped with an `AwsChunkDecodingInputStream`,
which makes no sense because the file input stream is never chunked/encoded.

This was not detected by tests before, because also in the integration tests
v4 signing is not used (because the related request header is set to "UNSIGNED-PAYLOAD").