Fixes #112: continuation token and start after
continuation token is now mapped to key in order to be able to get correct elements in case of removals/uploads.
start after was also fixed to work properly (esp. in case of missing key)
also continuation token and start after now share processing and follow Amazon's "This parameter is valid only in your first request. If the response is truncated, you can specify this parameter along with the continuation-token parameter, and then Amazon S3 ignores this parameter."