Fix accidental JDK9+ bytecode dependency
Source/Target parameters were deprecated in JDK8 and can cause errors
when running bytecode compiled in JDK9+ on JDKs lower than JDK9.
Use newly added `release` parameter instead.
The currently used identifier "1.8" is not allowed, it *has* to be "8".

The maven-javadoc-plugin tries to detect the Java version by retrieving
the "source" parameter of the compiler (D'OH!) so we have to tell it
to not detect the Java API version with "detectJavaApiLink=false".
Links to the JDK will default to the newest JDK version available in the
generated Javadoc now.

See #243 where @HackerTheMonkey analysed this issue pretty thoroughly,
and the official documentation:
https://docs.oracle.com/en/java/javase/11/migrate/index.html#JSMIG-GUID-77874D97-46F3-4DB5-85E4-2ACB5F8D760B

Fixes #243