Adds serialization tests for the DTO layer
This only adds tests and serialized responses to see what changes when
refactoring the code.

There are several issues with the current serialization:
Typos in fields. (e.g. "prefix" instead of "Prefix")
Superfluous fields (e.g. "truncated" or "part")
Some lists are not serialized correctly, as can be seen in
currently open issues.
Several types are not enforced, where fields in the POJOs are Strings
which should be Integer or enum.
Etag fields contain quoted values in all Amazon examples, while in
our serialization, they are all unquoted.
Date field serialization seems to be broken right now. Also, for dates,
we should be using "java.time" classes in favor of "java.util.Date".