Delete all files on JVM shutdown
All files that are created by a JVM need to be registered separately to
be deleted on exit. Simply registering the S3Mock root does not suffice.
Unfortunately, this only works when the JVM is shut down in an orderly
fashion.
If the JVM crashes, or if the Docker container crashes, there is no way
for us to delete the files uploaded to the S3Mock.

Fixes #249