Disable license header checks during maven release
With version 4.x, the license-maven-plugin checks the GIT history of a
file to determine when a file was created. This does not work on the GIT
clone the maven-release-plugin is creating just for the release.
We check the headers in the PRB _and_ in the "main" build after a PR is
merged. No need to check license headers again during a release.