Add Get- / Put- ObjectAcl
Implement GetObjectACL / PutObjectACL
Return / accept String instead of a POJO. We need to use JAX-B
annotations instead of Jackson annotations because AWS decided to use
xsi:type annotations in the XML representation, which are not supported
by Jackson. It doesn't seem to be possible to use bot JAX-B and Jackson
for (de-)serialization in parallel.

fixes #213 / #290