BucketController: Simplify bucket handling, error handling
Now that all objects are stored in their Bucket as a reference, it is
not necessary anymore to traverse directories and find all S3Objects.
If backing services (e.g. BucketStore) encapsulate all their internal
errors in certain exceptions (in this case IllegalStateExceptions), it
is possible to build a centralized error handler
(IllegalStateExceptionHandler) instead of handling errors separately in
every controller method.