Refactor into layers: Controller -> Service -> Store
Controllers work on Services, Services implement functionality on top of
Stores.
This is a "working version" of that pattern, with more to come - wording
of method parameters, JavaDoc updates/additions, tests, ...
Big missing feature: bucket access is not synchronized / guarded. If
multiple threads were to add or remove resources from/to a bucket, one
or more of the modifications will probably fail or interfere with each
other.