Correctly detect and use existing root folder
Forgot that this is a feature - mount the root folder into Docker, so
that we can look at the created files while the S3Mock is running, or,
if `retainFilesOnExit` is set to true, even after S3Mock shuts down.

I'll have to think about how to integration test this later.

Fixes #786