Spring Boot 3.0.0-M5 / Spring 6.0.0-M6 changes
JAX-B and Servlet APIs are now available in "jakarta" instead of "javax"
packages.
Classes with handler methods now have to be annotated with "@Controller"
which means that Spring will instantiate the class automatically, and
there's no way to fine-tune / manually instantiate the bean in a
configuration class. A ComponentScan with annotation filter is needed to
exclude these beans on startup...