Correctly detect EOL in chunked uploads
Even though the AWS documentation says that they always use "CRLF" as
the EOL indicator, it seems to depend on the system when using async
(native) clients.
Now, we use a BufferedReader to correctly detect and read full lines
where we need to.

MultiPartUploadV2IT#testMultipartUpload_asyncClient failed when using
a file >16KB before this fix, now it is successful.

Fixes #1818