Verify Copy / Object ITs
Handling for checksums and storageClass was surprising.
For some APIs, S3 does not return checksums unless specified in the
request. This is reflected in the tests, but not in the S3Mock.
The StorageClass "STANDARD" is the default storageClass and is not
returned by many APIs when requesting asset metadata.